+++ lib/disposable_cached_images.dart
@@ -1,27 +1,2 @@
-library disposable_cached_images;
-
-import 'dart:async';
-import 'dart:collection';
-import 'dart:typed_data';
-import 'dart:ui' as ui;
-
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import './helpers/interfaces.dart';
-import './image_info_data.dart';
-
-part './decoded_images/decoded_images.dart';
-part './download_progress.dart';
-part './image_decoder.dart';
-part './image_providers/arguments.dart';
-part './image_providers/base_provider.dart';
-part './image_providers/providers/assets.dart';
-part './image_providers/providers/local.dart';
-part './image_providers/providers/network.dart';
-part './image_providers/state.dart';
-part './image_providers/used_image_provider.dart';
-part './run_app_wrapper.dart';
-part './widget/main_widget.dart';
-part './widget/raw_image_widget.dart';
+export './disposable_cached_images_io.dart'
+    if (dart.library.html) 'disposable_cached_images_web/disposable_cached_images_web.dart';