+++ lib/disposable_cached_images_io.dart
@@ -1,0 +1,35 @@
+library disposable_cached_images_io;
+
+import 'dart:collection';
+import 'dart:convert';
+import 'dart:io';
+import 'dart:math' as math;
+import 'dart:ui' as ui;
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:isolate_generator_annotation/isolate_generator_annotation.dart';
+import 'package:path_provider/path_provider.dart';
+
+import './image_info_data.dart';
+
+part './decoded_images/decoded_images.dart';
+part './download_progress.dart';
+part './helpers/interfaces.dart';
+part './helpers/io/helper_io.dart';
+part './helpers/io/helper_io.g.dart';
+part './image_decoder.dart';
+part './image_providers/arguments.dart';
+part './image_providers/base_provider.dart';
+part './image_providers/providers/assets.dart';
+part './image_providers/providers/bytes.dart';
+part './image_providers/providers/local.dart';
+part './image_providers/providers/network.dart';
+part './image_providers/state.dart';
+part './image_providers/used_image_provider.dart';
+part './run_app_wrapper.dart';
+part './widget/main_widget.dart';
+part './widget/raw_image_widget.dart';