+++ lib/helpers/helper_stub.dart
@@ -1,5 +1,0 @@
-import 'interfaces.dart';
-
-HelperBase getInstance() {
-  throw UnsupportedError('unsupported platform');
-}